Mini: update best practices
