+++ io.openems.edge.fenecon.mini/src/io/openems/edge/fenecon/mini/gridmeter/FeneconMiniGridMeter.java
@@ -1,67 +1,14 @@
 package io.openems.edge.fenecon.mini.gridmeter;
 
-import org.osgi.service.cm.ConfigurationAdmin;
-import org.osgi.service.component.ComponentContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.ConfigurationPolicy;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.osgi.service.component.annotations.ReferencePolicy;
-import org.osgi.service.component.annotations.ReferencePolicyOption;
-import org.osgi.service.event.Event;
-import org.osgi.service.event.EventConstants;
 import org.osgi.service.event.EventHandler;
-import org.osgi.service.metatype.annotations.Designate;
 
 import io.openems.common.channel.Level;
-import io.openems.common.exceptions.OpenemsException;
-import io.openems.edge.bridge.modbus.api.AbstractOpenemsModbusComponent;
-import io.openems.edge.bridge.modbus.api.BridgeModbus;
-import io.openems.edge.bridge.modbus.api.ElementToChannelConverter;
-import io.openems.edge.bridge.modbus.api.ModbusProtocol;
-import io.openems.edge.bridge.modbus.api.element.BitsWordElement;
-import io.openems.edge.bridge.modbus.api.element.SignedWordElement;
-import io.openems.edge.bridge.modbus.api.task.FC3ReadRegistersTask;
 import io.openems.edge.common.channel.Doc;
 import io.openems.edge.common.component.OpenemsComponent;
-import io.openems.edge.common.event.EdgeEventConstants;
-import io.openems.edge.common.taskmanager.Priority;
-import io.openems.edge.fenecon.mini.FeneconMiniConstants;
-import io.openems.edge.meter.api.MeterType;
 import io.openems.edge.meter.api.SymmetricMeter;
-import io.openems.edge.timedata.api.Timedata;
 import io.openems.edge.timedata.api.TimedataProvider;
-import io.openems.edge.timedata.api.utils.CalculateEnergyFromPower;
 
-@Designate(ocd = Config.class, factory = true)
-@Component(//
-		name = "Fenecon.Mini.GridMeter", //
-		immediate = true, //
-		configurationPolicy = ConfigurationPolicy.REQUIRE, //
-		property = { //
-				EventConstants.EVENT_TOPIC + "=" + EdgeEventConstants.TOPIC_CYCLE_AFTER_PROCESS_IMAGE, //
-				"type=GRID" //
-		})
-public class FeneconMiniGridMeter extends AbstractOpenemsModbusComponent
-		implements SymmetricMeter, OpenemsComponent, TimedataProvider, EventHandler {
-
-	@Reference
-	protected ConfigurationAdmin cm;
-
-	@Reference(policy = ReferencePolicy.DYNAMIC, policyOption = ReferencePolicyOption.GREEDY, cardinality = ReferenceCardinality.OPTIONAL)
-	private volatile Timedata timedata = null;
-
-	private final CalculateEnergyFromPower calculateProductionEnergy = new CalculateEnergyFromPower(this,
-			SymmetricMeter.ChannelId.ACTIVE_PRODUCTION_ENERGY);
-	private final CalculateEnergyFromPower calculateConsumptionEnergy = new CalculateEnergyFromPower(this,
-			SymmetricMeter.ChannelId.ACTIVE_CONSUMPTION_ENERGY);
-
-	@Reference(policy = ReferencePolicy.STATIC, policyOption = ReferencePolicyOption.GREEDY, cardinality = ReferenceCardinality.MANDATORY)
-	protected void setModbus(BridgeModbus modbus) {
-		super.setModbus(modbus);
-	}
+public interface FeneconMiniGridMeter extends SymmetricMeter, OpenemsComponent, TimedataProvider, EventHandler {
 
 	public enum ChannelId implements io.openems.edge.common.channel.ChannelId {
 		COMMUNICATION_BREAK(Doc.of(Level.INFO) //
@@ -77,97 +24,6 @@ public class FeneconMiniGridMeter extends AbstractOpenemsModbusComponent
 		public Doc doc() {
 			return this.doc;
 		}
-	}
-
-	public FeneconMiniGridMeter() {
-		super(//
-				OpenemsComponent.ChannelId.values(), //
-				SymmetricMeter.ChannelId.values(), //
-				ChannelId.values() //
-		);
-	}
-
-	@Activate
-	void activate(ComponentContext context, Config config) throws OpenemsException {
-		if (super.activate(context, config.id(), config.alias(), config.enabled(), FeneconMiniConstants.UNIT_ID,
-				this.cm, "Modbus", config.modbus_id())) {
-			return;
-		}
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		super.deactivate();
-	}
-
-	@Override
-	protected ModbusProtocol defineModbusProtocol() throws OpenemsException {
-		return new ModbusProtocol(this, //
-				new FC3ReadRegistersTask(4004, Priority.HIGH, //
-						m(SymmetricMeter.ChannelId.ACTIVE_POWER, new SignedWordElement(4004),
-								SIGNED_POWER_CONVERTER_AND_INVERT)), //
-				new FC3ReadRegistersTask(4811, Priority.LOW, //
-						m(new BitsWordElement(4811, this) //
-								.bit(13, ChannelId.COMMUNICATION_BREAK))) //
-		);
-	}
-
-	@Override
-	public MeterType getMeterType() {
-		return MeterType.GRID;
-	}
-
-	@Override
-	public String debugLog() {
-		return "L:" + this.getActivePower().asString();
-	}
-
-	private static final ElementToChannelConverter SIGNED_POWER_CONVERTER_AND_INVERT = new ElementToChannelConverter(//
-			value -> {
-				if (value == null) {
-					return null;
-				}
-				int intValue = (Short) value;
-				if (intValue == -10_000) {
-					return 0; // ignore '-10_000'
-				}
-				return intValue * -1; // invert
-			}, //
-			value -> value);
-
-	@Override
-	public void handleEvent(Event event) {
-		switch (event.getTopic()) {
-		case EdgeEventConstants.TOPIC_CYCLE_AFTER_PROCESS_IMAGE:
-			this.calculateEnergy();
-			break;
-		}
-	}
-
-	/**
-	 * Calculate the Energy values from ActivePower.
-	 */
-	private void calculateEnergy() {
-		// Calculate Energy
-		Integer activePower = this.getActivePower().get();
-		if (activePower == null) {
-			// Not available
-			this.calculateProductionEnergy.update(null);
-			this.calculateConsumptionEnergy.update(null);
-		} else if (activePower > 0) {
-			// Buy-From-Grid
-			this.calculateProductionEnergy.update(activePower);
-			this.calculateConsumptionEnergy.update(0);
-		} else {
-			// Sell-To-Grid
-			this.calculateProductionEnergy.update(0);
-			this.calculateConsumptionEnergy.update(activePower * -1);
-		}
-	}
-
-	@Override
-	public Timedata getTimedata() {
-		return this.timedata;
 	}
 
 }
